implemented search algo
